Updating the auth logic
