+++ JeMPI_Apps/JeMPI_UI/src/services/ApiClient.ts
@@ -53,10 +53,7 @@ export class ApiClient {
         if (['post', 'patch', 'put', 'delete'].indexOf(method || '') !== -1) {
           const csrfToken = getCookie('XSRF-TOKEN')
           if (csrfToken) {
-            request.headers = {
-              ...request.headers,
-              'X-XSRF-TOKEN': csrfToken
-            }
+            request.headers['X-XSRF-TOKEN'] = csrfToken
           }
         }
         return request
