+++ JeMPI_Apps/JeMPI_UI/tests/test.utils/mocks/enviroments/DockerBase.ts
@@ -1,4 +1,5 @@
 import Docker from 'dockerode'
+import { getTestEnvConfig } from './Utils'
 
 export abstract class DockerBase{
 
@@ -15,22 +16,9 @@ export abstract class DockerBase{
     abstract HasLoadedFunc(): () => Promise<any>;
 
     GetEnvConfig(){
-        return {
-            isDev: process.env.NODE_ENV !== 'production',
-            apiUrl: process.env.REACT_APP_JEMPI_BASE_API_HOST
-              ? `${process.env.REACT_APP_JEMPI_BASE_API_HOST}:${process.env.REACT_APP_JEMPI_BASE_API_PORT}`
-              : `${window.location.protocol}//${window.location.hostname}:${process.env.REACT_APP_JEMPI_BASE_API_PORT}`,
-            shouldMockBackend: process.env.REACT_APP_MOCK_BACKEND === 'true',
-            KeyCloakUrl: process.env.KC_FRONTEND_URL || 'http://localhost:9088',
-            KeyCloakRealm: process.env.KC_REALM_NAME || 'platform-realm',
-            KeyCloakClientId: process.env.KC_JEMPI_CLIENT_ID || 'jempi-oauth',
-            useSso: process.env.REACT_APP_ENABLE_SSO === 'true',
-            maxUploadCsvSize: +(
-              process.env.REACT_APP_MAX_UPLOAD_CSV_SIZE_IN_MEGABYTES || 128
-            ),
-            showBrandLogo: process.env.REACT_APP_SHOW_BRAND_LOGO === 'true'
-          } 
+        return getTestEnvConfig()
     }
+    
     async Start(){
         await this.CreateService()
     }
