+++ JeMPI_Apps/JeMPI_UI/tests/test.utils/mocks/enviroments/MockKeyCloak/MockKeyCloak.ts
@@ -3,11 +3,10 @@ import axios from 'axios'
 import Docker from 'dockerode'
 import path from 'path';
 import fs from 'fs';
-import { config } from '../../../../../src/config'
+import { DockerBase } from '../DockerBase';
 
-class MockKeyCloack {
-    docker:Docker
-    serviceName:string
+class MockKeyCloack extends DockerBase{
+
     keyClockUrl:string
     keyClockUrlPort:string
     keyClockRealm:string
@@ -15,6 +14,8 @@ class MockKeyCloack {
     addtionalUser:any[]
 
     constructor(userConfig:any[]=[]){
+        super()
+        const config = this.GetEnvConfig();
         const parsedUrl = new URL(config.KeyCloakUrl);
 
         this.keyClockUrl = config.KeyCloakUrl
@@ -23,7 +24,6 @@ class MockKeyCloack {
         this.keyClockClient = config.KeyCloakClientId
 
         this.docker = new Docker()
-        this.serviceName = "JeMPIMockKeyCloak"
         this.addtionalUser = userConfig    
     }
 
@@ -121,56 +121,14 @@ class MockKeyCloack {
           }
     }
 
-    async CreateService(){
-        const waitToLoad = (checkPromise:() => Promise<any>, waitFor: 'resolve' | 'reject') => {
-            return new Promise( (resolve:any, reject:any) => {
-                let time= 0
-                const addTime = () =>{
-                    if (time > 60000){
-                        reject("Timeout starting keycloak")
-                    }
-                    time += 1000
-                }
-                
-                const interval = setInterval(() => checkPromise().then(() => {
-                    if (waitFor === 'resolve'){
-                        clearInterval(interval)
-                        resolve()
-                    }else{
-                        addTime()
-                    }
-                    
-                }).catch(() => {
-                    if (waitFor === 'reject'){
-                        clearInterval(interval)
-                        resolve()
-                    }else{
-                        addTime()
-                    }
-                }), 1000)
-            })
-        }
-
-        const containers = await this.docker.listContainers()
-        const targetContainerInfo = containers.find(container => container.Names.includes("/"+this.serviceName));
-        if (targetContainerInfo){
-            const targetContainer = await this.docker.getContainer(targetContainerInfo.Id)
-            await targetContainer.stop()
-            const stream = await targetContainer.attach({ stream: true, stdout: true, stderr: true })
-            stream.pipe(process.stdout)
-            await waitToLoad(() => this.docker.getContainer(targetContainerInfo.Id).inspect(), 'reject')
-            
-
-        }
-
-        const container = await this.docker.createContainer(this.GetConfig())
-        await container.start()
-        const stream = await container.attach({ stream: true, stdout: true, stderr: true })
-        stream.pipe(process.stdout)
-        await waitToLoad(() => axios.get(this.keyClockUrl), 'resolve')
-          
-
+    GetServiceName(): string {
+        return "JeMPIMockKeyCloak"
     }
+
+    HasLoadedFunc(): () => Promise<any> {
+        return () => axios.get(this.keyClockUrl)
+    }
+    
 }
 
 let fullPath:string = ""
